refactor: move tmp helper to src
