+++ package.json
@@ -66,6 +66,7 @@
     "minimatch": "^3.0.4",
     "semver": "7.3.4",
     "standard-version": "9.1.1",
+    "tmp": "0.2.1",
     "yargs": "^16.0.0"
   },
   "devDependencies": {
@@ -86,7 +87,6 @@
     "renovate-config-standard": "^2.0.0",
     "sinon": "^9.0.0",
     "sinon-chai": "^3.3.0",
-    "tmp": "0.2.1",
     "yargs-help-output": "^1.0.4"
   }
 }
