+++ test/build-dep-graph-test.js
@@ -4,7 +4,7 @@ const { describe, it, setUpSinon } = require('./helpers/mocha');
 const { expect } = require('./helpers/chai');
 const path = require('path');
 const buildDepGraph = require('../src/build-dep-graph');
-const { createTmpDir } = require('./helpers/tmp');
+const { createTmpDir } = require('../src/tmp');
 const fixturify = require('fixturify');
 const stringifyJson = require('../src/json').stringify;
 const { matchPath } = require('./helpers/matchers');
